+++ Ambassador/src/utopia/ambassador/database/access/single/organization/DbTask.scala
@@ -1,15 +1,24 @@
 package utopia.ambassador.database.access.single.organization
 
 import utopia.ambassador.database.access.many.scope.DbScopes
+import utopia.citadel.database.Tables
 import utopia.citadel.database.access.many.description.DbDescriptions
+import utopia.flow.generic.ValueConversions._
+import utopia.vault.nosql.template.Indexed
+import utopia.vault.nosql.view.{SubView, UnconditionalView}
 
 /**
   * Used for accessing individual tasks' data in the DB
   * @author Mikko Hilpinen
   * @since 11.7.2021, v1.0
   */
-object DbTask
+object DbTask extends UnconditionalView with Indexed
 {
+	// IMPLEMENTED  ------------------------
+	
+	override def table = Tables.task
+	
+	
 	// OTHER    ----------------------------
 	
 	/**
@@ -21,8 +30,10 @@ object DbTask
 	
 	// NESTED   ----------------------------
 	
-	class DbSingleTask(val taskId: Int)
+	class DbSingleTask(val taskId: Int) extends SubView
 	{
+		// COMPUTED -------------------------
+		
 		/**
 		  * @return An access point to this task's descriptions
 		  */
@@ -32,5 +43,12 @@ object DbTask
 		  * @return An access point to this task's scopes
 		  */
 		def scopes = DbScopes.forTaskWithId(taskId)
+		
+		
+		// IMPLEMENTED  --------------------
+		
+		override protected def parent = DbTask
+		
+		override def filterCondition = index <=> taskId
 	}
 }