+++ Citadel/src/utopia/citadel/database/access/id/many/DbUserRoleIds.scala
@@ -4,7 +4,7 @@ import utopia.citadel.database.Tables
 import utopia.citadel.database.model.organization.RoleRightModel
 import utopia.flow.generic.ValueConversions._
 import utopia.vault.database.Connection
-import utopia.vault.nosql.access.many.id.ManyIntIdAccess
+import utopia.vault.nosql.access.many.column.ManyIntIdAccess
 import utopia.vault.sql.{SelectDistinct, Where}
 import utopia.vault.sql.SqlExtensions._