+++ Exodus/src/utopia/exodus/database/access/id/DbLanguageIds.scala
@@ -2,7 +2,7 @@ package utopia.exodus.database.access.id
 
 import utopia.exodus.database.factory.language.LanguageFactory
 import utopia.flow.datastructure.immutable.Value
-import utopia.vault.nosql.access.many.id.ManyIdAccess
+import utopia.vault.nosql.access.many.column.ManyIdAccess
 
 /**
   * An access point to multiple language ids at once